feat(#562): fix page list of declarations
